requirements.txt
facebook-sdk == 2.0.0
tweepy == 3.5.0